refactor(cli): daemon commands move to src/cli/daemon.ts
